Move new traits in a separate file
