+++ zk_ee/src/common_traits/mod.rs
@@ -1,1 +1,2 @@
 pub mod key_like_with_bounds;
+pub mod usize_serializable_dynamic;
